Move Super Draw's layout toggle + visual settings into ToolControlsScaffold
Per Dylan's PR #691 review: in the stacked-panels layout, the toggle and
Visual settings button now sit at the bottom of ToolControlsScaffold's
right column (KeyOptionToggles), after Disallow paint-over, instead of
a sticky header above the stacked panels — condenses vertical space as
suggested. Extracted as SidebarLayoutToggle, shared with the tabbed
layout's tab-strip placement (DataCards).

Also swaps in the real DisallowPaintOver control here, replacing the
inert placeholder left over from before PR #689 landed.
